Legacy app Cleanup (#279)
* Cleanup

* Add back daos fetch
